# [struct_pack] enhance serialize performance by uninitialized resize