<!-- Thank you for your contribution! -->

## Why

In old version, if struct_pack save binary to a container, it will call member function `resize()`. which cause useless initialization.

## What is changing

Add a helper function `struct_pack::detail::resize()`, which will resize some container(std::basic_string<char/unsigned char/signed char>, std::vector<char/unsigned char/signed char/std::byte>) without initialization.

Now struct_pack::serialize is faster:

old version:
```
struct_pack serialize 1 rect :            3 ns
struct_pack serialize 20 rects :          17 ns
struct_pack serialize 1 person :          11 ns
struct_pack serialize 20 persons :        110 ns
struct_pack serialize 1 monster :         76 ns
struct_pack serialize 20 monsters :       652 ns
```
new version:
```
struct_pack serialize 1 rect :            2 ns
struct_pack serialize 20 rects :          11 ns
struct_pack serialize 1 person :          7 ns
struct_pack serialize 20 persons :        98 ns
struct_pack serialize 1 monster :         76 ns
struct_pack serialize 20 monsters :       618 ns
```

## Example